Atualizado validações de ingrediente

+++ src/main/java/com/mensal/sliceCtrl/service/IngredienteService.java
@@ -48,11 +48,23 @@ public class IngredienteService {
 
 
     public Ingredientes cadastrar(IngredientesDTO ingredientesDTO) {
+        Ingredientes ingredientesBanco = this.ingredienteRepository.findByNome(ingredientesDTO.getNomeIngrediente());
+        Assert.isTrue(ingredientesBanco == null, "Já possui um ingrediente com esse nome!");
+
         Ingredientes ingredientes = toIngredientes(ingredientesDTO);
         return this.ingredienteRepository.save(ingredientes);
     }
 
     public Ingredientes editar(IngredientesDTO ingredientesDTO, Long id) {
+
+        Ingredientes ingredientesBanco = this.ingredienteRepository.findByNome(ingredientesDTO.getNomeIngrediente());
+
+        if(ingredientesBanco != null){
+
+        if(ingredientesDTO.getId() != ingredientesBanco.getId()) {
+            Assert.isTrue(ingredientesBanco == null, "Já possui um ingrediente com esse nome!");
+        }
+        }
 
         final Ingredientes ingredientesBD = this.ingredienteRepository.findById(id).orElse(null);
         Assert.notNull(ingredientesBD, "Endereco inexistente!");
@@ -65,8 +77,18 @@ public class IngredienteService {
 
     public void delete(final Long id) {
         final Ingredientes ingredientes = this.ingredienteRepository.findById(id).orElse(null);
-        Assert.notNull(ingredientes, "Endereco inexistente!");
-        this.ingredienteRepository.delete(ingredientes);
+        Assert.notNull(ingredientes, "Ingrediente inexiste!");
+
+
+        if (ingredientes != null) {
+            if (!ingredientes.getSabores().isEmpty()) {
+                throw new IllegalArgumentException("Não é possível excluir o ingrediente devido às relações com sabores existentes.");
+            } else {
+                this.ingredienteRepository.delete(ingredientes);
+            }
+            }
+
     }
 
 }
+
